+++ samples/sampler/stories/qa/Scroller.js
@@ -261,6 +261,7 @@ export const EditableList = (args) => {
 			hoverToScroll={args['hoverToScroll']}
 			key={args['scrollMode']}
 			noScrollByWheel={args['noScrollByWheel']}
+			onClick={action('onClickScroller')}
 			onKeyDown={action('onKeyDown')}
 			onScrollStart={action('onScrollStart')}
 			onScrollStop={action('onScrollStop')}
@@ -278,6 +279,7 @@ export const EditableList = (args) => {
 							<ImageItem
 								src={item.src}
 								className={css.imageItem}
+								onClick={action('onClickItem')}
 							>
 								{`Image ${item.index}`}
 							</ImageItem>